Update ML-SUPERB2 MMS CTC recipe
