+++ egs2/ml_superb2/asr1/conf/train_asr.yaml
@@ -1,1 +1,1 @@
-tuning/train_mms_baseline.yaml
+tuning/train_mms_e_branchformer.yaml
