--- egs2/ml_superb2/asr1/conf/tuning/train_mms_ctc_e_branchformer.yaml
@@ -1,83 +1,0 @@
-encoder: e_branchformer
-encoder_conf:
-    output_size: 512
-    attention_heads: 8
-    attention_layer_type: rel_selfattn
-    pos_enc_layer_type: rel_pos
-    rel_pos_type: latest
-    cgmlp_linear_units: 3072
-    cgmlp_conv_kernel: 31
-    use_linear_after_conv: false
-    gate_activation: identity
-    num_blocks: 12
-    dropout_rate: 0.1
-    positional_dropout_rate: 0.1
-    attention_dropout_rate: 0.1
-    input_layer: conv2d2
-    layer_drop_rate: 0.0
-    linear_units: 2048
-    positionwise_layer_type: linear
-    use_ffn: true
-    macaron_ffn: false
-    merge_conv_kernel: 31
-
-model_conf:
-    ctc_weight: 1.0
-    extract_feats_in_collect_stats: false
-
-use_amp: true
-unused_parameters: false
-freeze_param: ["frontend.upstream"]
-num_att_plot: 0
-
-frontend: s3prl
-frontend_conf:
-    frontend_conf:
-        upstream: hf_wav2vec2_custom
-        path_or_url: facebook/mms-1b
-    download_dir: ./hub
-    multilayer_feature: True
-
-preencoder: linear
-preencoder_conf:
-    input_size: 1280
-    output_size: 80
-
-num_workers: 4
-batch_type: sorted
-batch_size: 8
-accum_grad: 4
-patience: none
-init: none
-best_model_criterion:
--   - valid
-    - loss
-    - min
-keep_nbest_models: 2
-
-optim: adam
-optim_conf:
-    lr: 0.0001
-    weight_decay: 0.000001
-scheduler: warmuplr
-scheduler_conf:
-    warmup_steps: 25000
-
-specaug: specaug
-specaug_conf:
-    apply_time_warp: true
-    time_warp_window: 5
-    time_warp_mode: bicubic
-    apply_freq_mask: true
-    freq_mask_width_range:
-    - 0
-    - 27
-    num_freq_mask: 2
-    apply_time_mask: true
-    time_mask_width_ratio_range:
-    - 0.
-    - 0.05
-    num_time_mask: 10
-
-num_iters_per_epoch: 20000
-max_epoch: 20
